+++ cli/src/cli.rs
@@ -92,7 +92,6 @@ Path to the configuration file. Defaults to $XDG_CONFIG_HOME/aim/config.toml on 
                     .subcommand(CmdTodoList::command()),
             )
             .subcommand(CmdTodoDone::command())
-            .subcommand(CmdTodoUndo::command().hide(true)) // TODO: remove in v0.4.0
             .subcommand(CmdGenerateCompletion::command())
     }
 
@@ -136,7 +135,6 @@ Path to the configuration file. Defaults to $XDG_CONFIG_HOME/aim/config.toml on 
                 _ => unreachable!(),
             },
             Some((CmdTodoDone::NAME, matches)) => TodoDone(CmdTodoDone::from(matches)),
-            Some((CmdTodoUndo::NAME, matches)) => Undo(CmdTodoUndo::from(matches)),
             Some((CmdGenerateCompletion::NAME, matches)) => {
                 GenerateCompletion(CmdGenerateCompletion::from(matches))
             }
@@ -190,9 +188,6 @@ pub enum Commands {
     /// List todos
     TodoList(CmdTodoList),
 
-    /// Mark a todo as undone
-    Undo(CmdTodoUndo),
-
     /// Generate shell completion
     GenerateCompletion(CmdGenerateCompletion),
 }
@@ -214,13 +209,6 @@ impl Commands {
             TodoDone(a)  => Self::run_with(config, |x| a.run(x).boxed()).await,
             TodoUndo(a)  => Self::run_with(config, |x| a.run(x).boxed()).await,
             TodoList(a)  => Self::run_with(config, |x| a.run(x).boxed()).await,
-            Undo(a) => {
-                println!(
-                    "{} `aim undo` is now `aim todo undo`, the shortcut will be removed in v0.4.0",
-                    "Deprecated:".yellow(),
-                );
-                Self::run_with(config, |x| a.run(x).boxed()).await
-            },
             GenerateCompletion(a) => a.run(),
         }
     }
@@ -419,17 +407,6 @@ mod tests {
                 );
             }
             _ => panic!("Expected TodoDone command"),
-        }
-    }
-
-    #[test]
-    fn test_parse_undo() {
-        let cli = Cli::try_parse_from(vec!["test", "undo", "id1"]).unwrap();
-        match cli.command {
-            Commands::Undo(cmd) => {
-                assert_eq!(cmd.ids, vec![Id::ShortIdOrUid("id1".to_string())]);
-            }
-            _ => panic!("Expected Undo command"),
         }
     }